Fixed permissions behavior in hellofs.
The tests were recently broken on Linux when default_permissions was set
when mounting. On Linux (but not on OS X), the uid sending the Init
request is root, so all of the files appear to be owned by root. This
was fine until we told the kernel to check permissions.